working create db
